Move all database related functions into their own module
