+++ diesel_cli/src/main.rs
@@ -4,40 +4,22 @@ extern crate diesel;
 extern crate dotenv;
 
 mod database_error;
+#[macro_use]
+mod database;
 
 use chrono::*;
 use clap::{App, AppSettings, Arg, ArgMatches, SubCommand};
-use diesel::expression::sql;
-use diesel::types::Bool;
 use diesel::pg::PgConnection;
 use diesel::sqlite::SqliteConnection;
 use diesel::migrations::schema::*;
 use diesel::types::{FromSql, VarChar};
-use diesel::{migrations, Connection, select, LoadDsl, Insertable};
+use diesel::{migrations, Connection, Insertable};
 use std::error::Error;
 use std::io::stdout;
 use std::path::{PathBuf, Path};
 use std::{env, fs};
 
 use self::database_error::{DatabaseError, DatabaseResult};
-
-macro_rules! call_with_conn {
-    ( $database_url:ident,
-      $func:path
-    ) => {{
-        match backend(&$database_url) {
-            "postgres" => {
-                let conn = PgConnection::establish(&$database_url).unwrap();
-                $func(&conn)
-            },
-            "sqlite" => {
-                let conn = SqliteConnection::establish(&$database_url).unwrap();
-                $func(&conn)
-            },
-            _ => unreachable!("The backend function should ensure we never get here."),
-        }
-    }};
-}
 
 fn main() {
     let database_arg = || Arg::with_name("DATABASE_URL")
@@ -108,7 +90,7 @@ fn main() {
 }
 
 fn run_migration_command(matches: &ArgMatches) {
-    let database_url = database_url(matches);
+    let database_url = database::database_url(matches);
 
     match matches.subcommand() {
         ("run", Some(_)) => {
@@ -153,93 +135,15 @@ fn run_setup_command(matches: &ArgMatches) {
                         .map_err(handle_error).unwrap()
                        );
 
-    setup_database(matches).unwrap_or_else(handle_error);
+    database::setup_database(matches).unwrap_or_else(handle_error);
 }
 
 fn run_database_command(matches: &ArgMatches) {
     match matches.subcommand() {
-        ("setup", Some(args)) => setup_database(args).unwrap_or_else(handle_error),
-        ("reset", Some(args)) => reset_database(args).unwrap_or_else(handle_error),
+        ("setup", Some(args)) => database::setup_database(args).unwrap_or_else(handle_error),
+        ("reset", Some(args)) => database::reset_database(args).unwrap_or_else(handle_error),
         _ => unreachable!("The cli parser should prevent reaching here"),
     };
-}
-
-pub fn reset_database(args: &ArgMatches) -> DatabaseResult<()> {
-    try!(drop_database(&database_url(args)));
-    setup_database(args)
-}
-
-pub fn setup_database(args: &ArgMatches) -> DatabaseResult<()> {
-    let database_url = database_url(args);
-
-    try!(create_database_if_needed(&database_url));
-    create_schema_table_and_run_migrations_if_needed(&database_url)
-}
-
-/// Creates the __diesel_schema_migrations table if it doesn't exist. If the
-/// table didn't exist, it also runs any pending migrations. Returns a
-/// `DatabaseError::ConnectionError` if it can't create the table, and exits
-/// with a migration error if it can't run migrations.
-fn create_schema_table_and_run_migrations_if_needed(database_url: &String)
-    -> DatabaseResult<()>
-{
-    if !schema_table_exists(database_url).map_err(handle_error).unwrap() {
-        try!(call_with_conn!(database_url, migrations::create_schema_migrations_table_if_needed));
-        call_with_conn!(database_url, migrations::run_pending_migrations).unwrap_or_else(handle_error);
-    };
-    Ok(())
-}
-
-/// Drops the database specified in the connection url. It returns an error
-/// if it was unable to drop the database.
-fn drop_database(database_url: &String) -> DatabaseResult<()> {
-    match backend(database_url) {
-        "postgres" => {
-            let (database, postgres_url) = split_pg_connection_string(database_url);
-            println!("Dropping database: {}", database);
-            let conn = try!(PgConnection::establish(&postgres_url));
-            try!(conn.silence_notices(|| {
-                conn.execute(&format!("DROP DATABASE IF EXISTS {}", database))
-            }));
-        },
-        "sqlite" => {
-            println!("Dropping database: {}", database_url);
-            try!(fs::remove_file(&database_url));
-        },
-        _ => unreachable!("The backend function should ensure we never get here."),
-    }
-    Ok(())
-}
-
-/// Creates the database specified in the connection url. It returns an error
-/// it it was unable to create the database.
-fn create_database_if_needed(database_url: &String)
-    -> DatabaseResult<()>
-{
-    match backend(database_url) {
-        "postgres" => {
-            if PgConnection::establish(database_url).is_err() {
-                let(database, postgres_url) = split_pg_connection_string(database_url);
-                try!(create_postgres_database(&postgres_url, &database));
-            }
-        },
-        "sqlite" => {
-            println!("Creating database: {}", database_url);
-            try!(SqliteConnection::establish(database_url));
-        },
-        _ => unreachable!("The backend function should ensure we never get here."),
-    }
-
-    Ok(())
-}
-
-fn create_postgres_database(database_url: &String, database: &String)
-    -> DatabaseResult<()>
-{
-    let conn = try!(PgConnection::establish(database_url));
-    println!("Creating database: {}", database);
-    try!(conn.execute(&format!("CREATE DATABASE {}", database)));
-    Ok(())
 }
 
 /// Looks for a migrations directory in the current path and all parent paths,
@@ -260,13 +164,6 @@ fn find_project_root() -> DatabaseResult<PathBuf> {
     search_for_cargo_toml_directory(&try!(env::current_dir()))
 }
 
-fn split_pg_connection_string(database_url: &String) -> (String, String) {
-    let mut split: Vec<&str> = database_url.split("/").collect();
-    let database = split.pop().unwrap();
-    let postgres_url = split.join("/");
-    (database.to_owned(), postgres_url)
-}
-
 /// Searches for the directory that holds the project's Cargo.toml, and returns
 /// the path if it found it, or returns a `DatabaseError::CargoTomlNotFound`.
 fn search_for_cargo_toml_directory(path: &Path) -> DatabaseResult<PathBuf> {
@@ -277,32 +174,6 @@ fn search_for_cargo_toml_directory(path: &Path) -> DatabaseResult<PathBuf> {
         path.parent().map(search_for_cargo_toml_directory)
             .unwrap_or(Err(DatabaseError::CargoTomlNotFound))
     }
-}
-
-/// Returns true if the '__diesel_schema_migrations' table exists in the
-/// database we connect to, returns false if it does not.
-pub fn schema_table_exists(database_url: &String) -> DatabaseResult<bool> {
-    let result = match backend(database_url) {
-        "postgres" => {
-            let conn = PgConnection::establish(database_url).unwrap();
-            try!(select(sql::<Bool>("EXISTS \
-                    (SELECT 1 \
-                     FROM information_schema.tables \
-                     WHERE table_name = '__diesel_schema_migrations')"))
-                .get_result(&conn))
-        },
-        "sqlite" => {
-            let conn = SqliteConnection::establish(database_url).unwrap();
-            try!(select(sql::<Bool>("EXISTS \
-                    (SELECT 1 \
-                     FROM sqlite_master \
-                     WHERE type = 'table' \
-                     AND name = '__diesel_schema_migrations')"))
-                .get_result(&conn))
-        },
-        _ => unreachable!("The backend function should ensure we never get here."),
-    };
-    Ok(result)
 }
 
 /// Reverts the most recent migration, and then runs it again, all in a
@@ -321,26 +192,6 @@ fn redo_latest_migration<Conn>(conn: &Conn) where
 
 fn handle_error<E: Error>(error: E) {
     panic!("{}", error);
-}
-
-pub fn database_url(matches: &ArgMatches) -> String {
-    dotenv::dotenv().ok();
-
-    matches.value_of("DATABASE_URL")
-        .map(|s| s.into())
-        .or(env::var("DATABASE_URL").ok())
-        .expect("The --database-url argument must be passed, \
-                or the DATABASE_URL environment variable must be set.")
-}
-
-/// Returns a &str representing the type of backend being used, determined
-/// by the format of the database url.
-pub fn backend(database_url: &String) -> &str {
-    if database_url.starts_with("postgres://") || database_url.starts_with("postgresql://") {
-        "postgres"
-    } else {
-        "sqlite"
-    }
 }
 
 // Converts an absolute path to a relative path, with the restriction that the
@@ -393,149 +244,15 @@ where I: Iterator<Item = A> + Clone,
 mod tests {
     extern crate tempdir;
 
-    use dotenv::dotenv;
-    use diesel::Connection;
-
     use database_error::DatabaseError;
 
     use self::tempdir::TempDir;
 
-    use std::{env, fs};
+    use std::fs;
     use std::path::PathBuf;
 
     use super::convert_absolute_path_to_relative;
     use super::search_for_cargo_toml_directory;
-    use super::{create_database_if_needed, drop_database, schema_table_exists};
-    use super::split_pg_connection_string;
-
-    #[cfg(feature = "postgres")]
-    type TestConnection = ::diesel::pg::PgConnection;
-    #[cfg(feature = "sqlite")]
-    type TestConnection = ::diesel::sqlite::SqliteConnection;
-
-    type TestBackend = <TestConnection as Connection>::Backend;
-
-    #[cfg(feature = "postgres")]
-    fn database_url(identifier: &str) -> String {
-        dotenv().ok();
-        let database_url = env::var("DATABASE_URL")
-            .expect("DATABASE_URL must be set in order to run diesel_cli tests");
-        let (_, base_url) = split_pg_connection_string(&database_url);
-        format!("{}/{}", base_url, identifier)
-    }
-
-    #[cfg(feature = "postgres")]
-    fn connection(database_url: &String) -> TestConnection {
-        ::diesel::pg::PgConnection::establish(database_url).unwrap()
-    }
-
-    #[cfg(feature = "postgres")]
-    fn teardown(_: String) {
-    }
-
-    #[cfg(feature = "sqlite")]
-    fn database_url(identifier: &str) -> String {
-        let dir = env::current_dir().unwrap();
-        let db_file = dir.join(identifier);
-        fs::remove_file(&db_file).ok();
-        fs::File::create(&db_file).unwrap();
-        db_file.to_str().unwrap().to_owned()
-    }
-
-    #[cfg(feature = "sqlite")]
-    fn connection(database_url: &String) -> TestConnection {
-        ::diesel::sqlite::SqliteConnection::establish(database_url).unwrap()
-    }
-
-    #[cfg(feature = "sqlite")]
-    fn teardown(database_path: String) {
-        fs::remove_file(&database_path).unwrap();
-    }
-
-    #[test]
-    fn schema_table_exists_finds_table() {
-        let database_url = database_url("test1");
-        create_database_if_needed(&database_url)
-            .expect("Unable to create test database");
-        let connection = connection(&database_url);
-        connection.silence_notices(|| {
-            connection.execute("DROP TABLE IF EXISTS __diesel_schema_migrations").unwrap();
-            connection.execute("CREATE TABLE __diesel_schema_migrations (version INTEGER)").unwrap();
-        });
-
-        assert!(schema_table_exists(&database_url).unwrap());
-
-        teardown(database_url);
-    }
-
-    #[test]
-    fn schema_table_exists_doesnt_find_table() {
-        let database_url = database_url("test2");
-        create_database_if_needed(&database_url)
-            .expect("Unable to create test database");
-        let connection = connection(&database_url);
-        connection.silence_notices(|| {
-            connection.execute("DROP TABLE IF EXISTS __diesel_schema_migrations").unwrap();
-        });
-
-        assert!(!schema_table_exists(&database_url).unwrap());
-
-        teardown(database_url);
-    }
-
-    #[test]
-    fn create_database_creates_the_database() {
-        let database_url = database_url("test3");
-        drop_database(&database_url).unwrap();
-        create_database_if_needed(&database_url).unwrap();
-        assert!(TestConnection::establish(&database_url).is_ok());
-
-        teardown(database_url);
-    }
-
-    #[test]
-    #[cfg(feature = "postgres")]
-    fn drop_database_drops_the_database() {
-        let database_url = database_url("test4");
-        drop_database(&database_url).unwrap();
-        assert!(TestConnection::establish(&database_url).is_err());
-    }
-
-    #[test]
-    #[cfg(feature = "sqlite")]
-    fn drop_database_drops_the_database() {
-        let database_url = database_url("test4");
-        drop_database(&database_url).unwrap();
-        assert!(fs::File::open(database_url).is_err());
-    }
-
-    // #[test]
-    // #[should_panic] // FIXME: Our migration structure is non-standard
-    // // we need to test this against a clean env with a normal structure
-    // // once we get integration test coverage (we can't change cwd in the test
-    // // process)
-    // fn create_schema_table_creates_diesel_table() {
-    //     let database_url = database_url("test5");
-    //     create_database_if_needed(&database_url)
-    //         .expect("Unable to create test database");
-    //     let connection = connection(&database_url);
-    //     connection.silence_notices(|| {
-    //         connection.execute("DROP TABLE IF EXISTS __diesel_schema_migrations").unwrap();
-    //     });
-    //     assert!(!schema_table_exists(&database_url).unwrap());
-    //     create_schema_table_and_run_migrations_if_needed(&database_url).unwrap();
-    //     assert!(schema_table_exists(&database_url).unwrap());
-    //
-    //     teardown(database_url);
-    // }
-
-    #[test]
-    fn split_pg_connection_string_returns_postgres_url_and_database() {
-        let database = "database".to_owned();
-        let postgres_url = "postgresql://localhost:5432".to_owned();
-        let database_url = format!("{}/{}", postgres_url, database);
-        assert_eq!((database, postgres_url), split_pg_connection_string(&database_url));
-    }
 
     #[test]
     fn toml_directory_find_cargo_toml() {
@@ -560,19 +277,19 @@ mod tests {
     #[test]
     fn convert_absolute_path_to_relative_works() {
         assert_eq!(PathBuf::from("migrations/12345_create_user"),
-                        convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
-                                                            &mut PathBuf::from("projects/foo")));
+                   convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
+                                                     &mut PathBuf::from("projects/foo")));
         assert_eq!(PathBuf::from("../migrations/12345_create_user"),
-                        convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
-                                                            &mut PathBuf::from("projects/foo/src")));
+                   convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
+                                                     &mut PathBuf::from("projects/foo/src")));
         assert_eq!(PathBuf::from("../../../migrations/12345_create_user"),
-                        convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
-                                                            &mut PathBuf::from("projects/foo/src/controllers/errors")));
+                   convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
+                                                     &mut PathBuf::from("projects/foo/src/controllers/errors")));
         assert_eq!(PathBuf::from("12345_create_user"),
-                        convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
-                                                            &mut PathBuf::from("projects/foo/migrations")));
+                   convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
+                                                     &mut PathBuf::from("projects/foo/migrations")));
         assert_eq!(PathBuf::from("../12345_create_user"),
-                        convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
-                                                            &mut PathBuf::from("projects/foo/migrations/67890_create_post")));
+                   convert_absolute_path_to_relative(&mut PathBuf::from("projects/foo/migrations/12345_create_user"),
+                                                     &mut PathBuf::from("projects/foo/migrations/67890_create_post")));
     }
 }
